+++ docs/ReleaseNotes.md
@@ -1,5 +1,12 @@
 // Before any commit, use the command line: npx tsc --noEmit
 
-# 1.9.28 Missing Media Status Feedback
+# 1.9.29 Image download
 
-- Now project buttons display a status of "Media no uploaded" if the project uploads, but there is media not uploaded yet.
+- No more possible do upload a image that does not exist on user device.
+- Added a new method "identifyMissingPicturesOnCache" on Cache Service. Responsible to identify pictures not available on user device.
+- New Alert Local component abstracted: <CredentialsDisplay />
+- New Alert Local component abstracted: <ErrorDisplay />
+- New Alert Local component abstracted: <TemplateWidgetCopyDisplay />
+- Removed the method changeAllIds from Project Service.
+- Download Service downloadPictures method now breaks the fetch if the proccess is aborted.
+- Database Service Media Sync now considers media download sync.